#187: Refactor edit and read modes into own components
