+++ src/components/Dashboard/Profile/sections/AccompanyingDetails/AccompanyingDetailsDisplay.tsx
@@ -1,0 +1,62 @@
+import { EditableField } from "@/components/EditableField/EditableField";
+import { EMPTY_PLACEHOLDER_VALUE } from "@/config/constants";
+import { format } from "date-fns";
+import { useTranslation } from "react-i18next";
+import { AccompanyingDetailsFormData } from "./accompanyingDetailsSchema";
+import { DateFieldRow, Details } from "./styles";
+
+type Props = {
+  values: AccompanyingDetailsFormData;
+  languageLabel: string;
+};
+
+export const AccompanyingDetailsDisplay = ({ values, languageLabel }: Props) => {
+  const { t } = useTranslation();
+
+  return (
+    <Details data-testid="accompanying-details-display">
+      <EditableField
+        mode="display"
+        type="text"
+        label={t("dashboard.opportunityProfile.accompanyingDetails.appointmentAddress")}
+        value={values.appointmentAddress || ""}
+        setValue={() => {}}
+      />
+
+      <DateFieldRow data-testid="appointment-date-field">
+        <label>{t("dashboard.opportunityProfile.accompanyingDetails.appointmentDate")}</label>
+        <span>{values.appointmentDate ? format(values.appointmentDate, "dd.MM.yyyy") : EMPTY_PLACEHOLDER_VALUE}</span>
+      </DateFieldRow>
+
+      <DateFieldRow data-testid="appointment-time-field">
+        <label>{t("dashboard.opportunityProfile.accompanyingDetails.appointmentTime")}</label>
+        <span>{values.appointmentTime || EMPTY_PLACEHOLDER_VALUE}</span>
+      </DateFieldRow>
+
+      <EditableField
+        mode="display"
+        type="text"
+        label={t("dashboard.opportunityProfile.accompanyingDetails.refugeeNumber")}
+        value={values.refugeeNumber || ""}
+        setValue={() => {}}
+      />
+
+      <EditableField
+        mode="display"
+        type="text"
+        label={t("dashboard.opportunityProfile.accompanyingDetails.refugeeName")}
+        value={values.refugeeName || ""}
+        setValue={() => {}}
+      />
+
+      <EditableField
+        mode="display"
+        type="radio-list"
+        label={t("dashboard.opportunityProfile.accompanyingDetails.languageToTranslate")}
+        value={languageLabel}
+        setValue={() => {}}
+        options={[]}
+      />
+    </Details>
+  );
+};
